move VerifierFn to proof service factory
