+++ node/proof_service_factory.go
@@ -1,12 +1,36 @@
 package node
 
 import (
+	"errors"
+
 	"github.com/celestiaorg/go-fraud"
 	"github.com/celestiaorg/go-fraud/fraudserv"
 	"github.com/ipfs/go-datastore"
+	abci "github.com/tendermint/tendermint/abci/types"
+	proxy "github.com/tendermint/tendermint/proxy"
 
 	"github.com/rollkit/rollkit/p2p"
+	"github.com/rollkit/rollkit/types"
 )
+
+var VerifierFn = func(proxyApp proxy.AppConns) func(fraudProof fraud.Proof) (bool, error) {
+	return func(fraudProof fraud.Proof) (bool, error) {
+		stateFraudProof, ok := fraudProof.(*types.StateFraudProof)
+		if !ok {
+			return false, errors.New("unknown fraud proof")
+		}
+		resp, err := proxyApp.Consensus().VerifyFraudProofSync(
+			abci.RequestVerifyFraudProof{
+				FraudProof:           &stateFraudProof.FraudProof,
+				ExpectedValidAppHash: stateFraudProof.ExpectedValidAppHash,
+			},
+		)
+		if err != nil {
+			return false, err
+		}
+		return resp.Success, nil
+	}
+}
 
 type ProofServiceFactory struct {
 	client        *p2p.Client
